+++ src/test/scala/uk/co/bigbeeconsultants/http/cache/CacheTest.scala
@@ -39,7 +39,7 @@ class CacheTest extends FunSuite {
     for (method <- List(GET, HEAD)) {
       val request = Request(method, "http://localhost/stuff")
       for (status <- List(Status.S200_OK, Status.S203_NotAuthoritative, Status.S300_MultipleChoice, Status.S301_MovedPermanently, Status.S410_Gone)) {
-        val cache = new Cache()
+        val cache = new InMemoryCache()
         val response = Response(request, status, MediaType.TEXT_PLAIN, "OK")
         cache.store(response)
         assert(cache.size === 1)
@@ -51,7 +51,7 @@ class CacheTest extends FunSuite {
     for (method <- List(GET, HEAD)) {
       val request = Request(method, "http://localhost/stuff")
       val status = Status.S404_NotFound
-      val cache = new Cache(assume404Age = 100)
+      val cache = new InMemoryCache(assume404Age = 100)
       val response = Response(request, status, MediaType.TEXT_PLAIN, "OK")
       cache.store(response)
       assert(cache.size === 1)
@@ -59,7 +59,7 @@ class CacheTest extends FunSuite {
   }
 
   test("store ignores other methods") {
-    val cache = new Cache()
+    val cache = new InMemoryCache()
     for (method <- List(POST, PUT, DELETE)) {
       val request = Request(method, "http://localhost/stuff")
       val response = Response(request, Status.S200_OK, MediaType.TEXT_PLAIN, "OK")
@@ -69,7 +69,7 @@ class CacheTest extends FunSuite {
   }
 
   test("store ignores no-cache and no-store cache-control values") {
-    val cache = new Cache()
+    val cache = new InMemoryCache()
     for (value <- List("no-cache", "no-store")) {
       val request = Request(GET, "http://localhost/stuff")
       val responseHeaders = Headers(CACHE_CONTROL -> value)